# Conflicts:
#	RouteAnalyzer/src/main/kotlin/it/polito/wa2/g20/Main.kt
#	RouteAnalyzer/src/main/kotlin/it/polito/wa2/g20/model/FuelEfficiency.kt
#	RouteAnalyzer/src/main/kotlin/it/polito/wa2/g20/service/DirectionPathCalculator.kt
#	RouteAnalyzer/src/main/kotlin/it/polito/wa2/g20/service/DistanceCalculator.kt
#	RouteAnalyzer/src/main/kotlin/it/polito/wa2/g20/service/FuelAnalyzer.kt
#	RouteAnalyzer/src/main/kotlin/it/polito/wa2/g20/service/RootTypeAnalyzer.kt

+++ RouteAnalyzer/src/main/kotlin/it/polito/wa2/g20/model/Waypoint.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class Waypoint(
-    val timestamp: Double,
+    val timestamp: Long,
     val latitude: Double,
     val longitude: Double
 )
